// packages/primitives/src/components/Theme/ThemeWorkbench.svelte.d.ts
import { type ThemeTokens } from '@equaltoai/greater-components-utils';
interface Props {
    initialColor?: string;
    onSave?: (theme: ThemeTokens) => void;
}
declare const ThemeWorkbench: import("svelte").Component<Props, {}, "">;
type ThemeWorkbench = ReturnType<typeof ThemeWorkbench>;
export default ThemeWorkbench;
//# sourceMappingURL=ThemeWorkbench.svelte.d.ts.map